+++ app/api/push-subscriptions/route.ts
@@ -1,4 +1,4 @@
-import {deleteSubscriptionFromDb, getSubscriptionsFromDb, saveSubscriptionToDb } from '@/app/utils/in-memory-db'
+import {deleteSubscriptionFromDb, getSubscriptionsFromDb, saveSubscriptionToDb } from '@/utils/in-memory-db'
 import {NextRequest, NextResponse } from 'next/server'
 import webPush, { PushSubscription } from 'web-push';
 
@@ -43,32 +43,6 @@ export async function DELETE(request: NextRequest) {
 
         const updatedDb = await deleteSubscriptionFromDb(subscriptionToDelete)
         return NextResponse.json({ message: 'Push subscription deleted', status: 200, updatedDb })
-    } catch(error) {
-        console.error(error);
-        throw NextResponse.json(
-            { error: 'Internal server error '},
-            {status: 500}
-        );
-    }
-}
-
-
-export async function GET(_: NextRequest) {
-    try{
-        const subscriptions = getSubscriptionsFromDb();
-        const notifications = subscriptions.map((subscription) => { //Noe som ikke funker her. Notification blir ikke sendt
-            const payload = JSON.stringify({
-                title: 'WebPush Notification!',
-                body: 'Hello World',
-            })
-            webPush.sendNotification(subscription, payload)
-        })
-
-        await Promise.all(notifications);
-
-        return NextResponse.json({
-            message: `${subscriptions.length} messages sent!`,
-        })
     } catch(error) {
         console.error(error);
         throw NextResponse.json(
