+++ public/manifest.json
@@ -1,11 +1,11 @@
 {
-  "name": "WebPush with Next.js",
-  "short_name": "WebPush",
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#B12A34",
-  "background_color": "#B12A34",
-  "description": "Example PWA with WebPush!",
+  "name": "TBD",
+  "short_name": "TBD",
+  "start_url": "TBD",
+  "display": "TBD",
+  "theme_color": "TBD",
+  "background_color": "TBD",
+  "description": "TBD",
   "icons": [],
   "related_applications": []
 }
