+++ utils/serviceWorker.ts
@@ -2,7 +2,7 @@ export const registerServiceWorker = async () => {
     if (!('serviceWorker' in navigator)) {
         throw Error("Service workers are not supported by this browser.");
     }
-    return navigator.serviceWorker.register('/service.js');
+    return navigator.serviceWorker.register('/service-worker.js');
 }
 export const getReadyServiceWorker = async () => {
     if (!('serviceWorker' in navigator)) {
